+++ src/nvim/regexp_nfa.c
@@ -6480,8 +6480,7 @@ static long nfa_regtry(nfa_regprog_T *prog,
 
         if (lpos->start != NULL && lpos->end != NULL)
           re_extmatch_out->matches[i] =
-            vim_strnsave(lpos->start,
-                (int)(lpos->end - lpos->start));
+            vim_strnsave(lpos->start, lpos->end - lpos->start);
       }
     }
   }
